% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing-led_arrange_outputs.R
\name{arrange_housing_led_outputs}
\alias{arrange_housing_led_outputs}
\title{Arrange ward-level housing-led outputs}
\usage{
arrange_housing_led_outputs(
  projection,
  population,
  births,
  deaths,
  in_migration,
  out_migration,
  fertility_rates,
  mortality_rates,
  in_migration_flows,
  out_migration_rates,
  first_proj_yr,
  last_proj_yr
)
}
\arguments{
\item{projection}{A list output from housing_led_core}

\item{population}{A dataframe of the population backseries}

\item{births}{A dataframe of the births backseries}

\item{deaths}{A dataframe of the deaths backseries}

\item{in_migration}{A dataframe of the in migration backseries}

\item{out_migration}{A dataframe of the out migration backseries}

\item{fertility_rates}{A dataframe of fertility rates}

\item{mortality_rates}{A dataframe of mortality rates}

\item{in_migration_flows}{A dataframe of input in migration flows}

\item{out_migration_rates}{A dataframe of input out migration rates}

\item{first_proj_yr}{Numeric. The first projection year}

\item{last_proj_yr}{Numeric. The last projection}
}
\value{
A list of projection outputs by component
}
\description{
Take the projection output list which is organised by year and rearrange it
so that its organised by component. Add on the backseries.
}
