% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing-led_loop.R
\name{housing_led_core}
\alias{housing_led_core}
\title{Project from a starting population using a housing-led methodology to
produce an output population and components}
\usage{
housing_led_core(
  start_population,
  trend_projection,
  communal_establishment_population,
  household_rep_rates,
  households,
  projection_year,
  ahs_mix
)
}
\arguments{
\item{start_population}{A dataframe. The starting population}

\item{trend_projection}{A list. The output from the small area trend model for
the year \code{projection_year}}

\item{communal_establishment_population}{A dataframe. The communal establishment
population}

\item{household_rep_rates}{A dataframe. Pre-calculated HHRs for model areas}

\item{households}{A dataframe. An input total household figure for model areas}

\item{projection_year}{Numeric. The year being projected}

\item{ahs_mix}{Numeric. A number between 0 and 1 defining the mix of AHS
from the HHR household model and the implied trend AHS}
}
\value{
A list of projected components
}
\description{
Project from a starting population using a housing-led methodology to
produce an output population and components
}
