% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing-led_outputs.R
\name{output_housing_led}
\alias{output_housing_led}
\title{Write model outputs to RDS and CSV}
\usage{
output_housing_led(
  projection,
  output_dir = config_list$output_dir,
  n_csv_elements = 7
)
}
\arguments{
\item{projection}{A list. The projection as processed through the arrange function}

\item{output_dir}{String. The path to the output folder where files will be written}

\item{n_csv_elements}{Numeric. The first n element of the projection list will be
written to csv. Default = 7.}
}
\description{
Outputs from the small area housing-led model, passed through the \code{arrange}
function are written out a RDS and CSV files
}
