% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_outputs.R
\name{output_projection}
\alias{output_projection}
\title{Write model outputs to RDS and CSV}
\usage{
output_projection(
  projection,
  output_dir = config_list$output_dir,
  n_csv_elements = 8
)
}
\arguments{
\item{projection}{A list. The projection as processed through the arrange function}

\item{output_dir}{String. The path to the output folder where files will be written}

\item{n_csv_elements}{Numeric. The first n element of the projection list will be
written to csv. Default = 8.}
}
\description{
Outputs from the small area trend model, passed through the \code{arrange}
function are written out a RDS and CSV files
}
